requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
passlib
flask-migrate
requests
redis
rq
jinja2